Extract DevTools.RuntimeDeploy.Engine class library
Move pure Domain/Infrastructure logic (BuildArtefactsScanner, FrameworkInfo,
FrameworkListManifestEditor, AssemblyProbe, FileLockInspector, etc.) out of
the WinExe DevTools.RuntimeDeploy project into a new headless
DevTools.RuntimeDeploy.Engine class library so it can be reused by a future
CLI tool. Types made public where needed for cross-assembly access.
