+++ src/DevTools.RuntimeDeploy/DevTools.RuntimeDeploy.csproj
@@ -16,7 +16,6 @@
 
   <ItemGroup>
     <PackageReference Include="CommunityToolkit.Mvvm" />
-    <PackageReference Include="NuGet.Versioning" />
     <PackageReference Include="WarpToolkit.ComponentModel" />
     <PackageReference Include="WarpToolkit.Desktop.AppServices" />
     <PackageReference Include="WarpToolkit.WinForms" />
@@ -24,6 +23,10 @@
     <PackageReference Include="WarpToolkit.WinForms.Async" />
     <PackageReference Include="WarpToolkit.WinForms.Extensions" />
     <PackageReference Include="WarpToolkit.WinForms.Specialized" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\DevTools.RuntimeDeploy.Engine\DevTools.RuntimeDeploy.Engine.csproj" />
   </ItemGroup>
 
   <ItemGroup>
