+++ src/DevTools.RuntimeDeploy/Infrastructure/CommandBatch.cs
@@ -1,4 +1,5 @@
 using System.Text;
+using DevTools.RuntimeDeploy.Engine.Infrastructure;
 
 namespace DevTools.RuntimeDeploy.Infrastructure;
