+++ src/DevTools.RuntimeDeploy/MainForm.cs
@@ -1,4 +1,4 @@
-using DevTools.RuntimeDeploy.Domain;
+using DevTools.RuntimeDeploy.Engine.Domain;
 using DevTools.RuntimeDeploy.Infrastructure;
 using DevTools.RuntimeDeploy.Views;
 using Microsoft.Extensions.DependencyInjection;
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Logging;
 using System.ComponentModel;
 using WarpToolkit.ComponentModel;
 using WarpToolkit.Desktop.AppServices;
-using static DevTools.RuntimeDeploy.Domain.BuildArtefactsScanner;
+using static DevTools.RuntimeDeploy.Engine.Domain.BuildArtefactsScanner;
 
 namespace DevTools.RuntimeDeploy;
