+++ src/DevTools.RuntimeDeploy/Views/DeployRuntimeView.cs
@@ -1,8 +1,8 @@
-using DevTools.RuntimeDeploy.Domain;
+using DevTools.RuntimeDeploy.Engine.Domain;
 using DevTools.RuntimeDeploy.Infrastructure;
 using Microsoft.Extensions.Logging;
 using System.Data;
-using static DevTools.RuntimeDeploy.Domain.BuildArtefactsScanner;
+using static DevTools.RuntimeDeploy.Engine.Domain.BuildArtefactsScanner;
 
 namespace DevTools.RuntimeDeploy.Views;
